✨ [#11] 회원 정보 조회 기능 추가 (#23)
* ✨ [#11] 회원 정보 조회 기능 추가

- 이메일, 캐시, 포인트, isAdmin 정보를 조회할 수 있습니다

* ✨ [#11] 빠트린 파일 추가

* 🐛 [#11] bug 수정

- 브랜치끼리 충돌 되는 부분 수정
- Member 엔티티 cash, point 타입 Integer 변경
- dto 폴더 request/response 으로 세분화 및 위치 이동
- MemberInfoDto true/false가 아닌 role 값으로 나오도록 수정

* ♻️ [#11] 코드 수정

- Optional으로 변경
- MemberDetailsDto 대신 Member 엔티티 사용으로 변경
- 예외처리 추가

* ♻️ [#11] 코드 수정

- 사용하지 않는 import 제거
- MemberDetailsDto 인증시 필요 없는 부분 일부 삭제

+++ goodsending/src/main/java/com/goodsending/member/entity/Member.java
@@ -1,7 +1,7 @@
 package com.goodsending.member.entity;
 
 import com.goodsending.global.entity.BaseEntity;
-import com.goodsending.member.dto.SignupRequestDto;
+import com.goodsending.member.dto.request.SignupRequestDto;
 import com.goodsending.member.type.MemberRole;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
@@ -30,10 +30,10 @@ public class Member extends BaseEntity {
   private String phoneNumber;
 
   @Column(name = "cash", nullable = true)
-  private Long cash;
+  private Integer cash;
 
   @Column(name = "point", nullable = true)
-  private Long point;
+  private Integer point;
 
   @Column(name = "code", nullable = true)
   private Long code;
